refactor: 함께해요 글쓰기 주소 체계 법정동으로 변경
