+++ app/src/main/java/com/umc/ttoklip/data/model/naver/reversegeocoding/AreaInfo.kt
@@ -1,8 +1,8 @@
-package kr.ac.tukorea.whereareu.data.model.naver
+package com.umc.ttoklip.data.model.naver.reversegeocoding
 
 data class AreaInfo(
     val name: String,
-    val coords:Coords
+    val coords: Coords
 )
 
 data class Coords(
