+++ app/src/main/java/com/umc/ttoklip/data/model/naver/reversegeocoding/Land.kt
@@ -1,5 +1,7 @@
 package kr.ac.tukorea.whereareu.data.model.naver
 
+import com.umc.ttoklip.data.model.naver.reversegeocoding.LandInfo
+
 data class Land(
     val number1: String,
     val number2: String,
