+++ app/src/main/java/com/umc/ttoklip/data/model/naver/reversegeocoding/ReverseGeocodingResponse.kt
@@ -1,8 +1,5 @@
 package com.umc.ttoklip.data.model.naver.reversegeocoding
 
-import kr.ac.tukorea.whereareu.data.model.naver.ReverseGeoCodingResult
-import kr.ac.tukorea.whereareu.data.model.naver.StatusResult
-
 data class ReverseGeocodingResponse(
     val status: StatusResult,
     val results: List<ReverseGeoCodingResult>
