+++ app/src/main/java/com/umc/ttoklip/data/model/naver/reversegeocoding/StatusResult.kt
@@ -1,4 +1,4 @@
-package kr.ac.tukorea.whereareu.data.model.naver
+package com.umc.ttoklip.data.model.naver.reversegeocoding
 
 data class StatusResult(
     val code: Int,
