+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/communication/CommunicationViewModelImpl.kt
@@ -32,8 +32,8 @@ class CommunicationViewModelImpl @Inject constructor(
     override val criteria: StateFlow<String>
         get() = _criteria
 
-    private val _streetInfo = MutableStateFlow("")
-    override val streetInfo: StateFlow<String>
+    private val _streetInfo = MutableStateFlow<List<String>>(emptyList())
+    override val streetInfo: StateFlow<List<String>>
         get() = _streetInfo
 
     override fun setCriteria(position: Int) {
@@ -59,7 +59,7 @@ class CommunicationViewModelImpl @Inject constructor(
     override fun getMemberStreetInfo() {
         viewModelScope.launch {
             mainTogethersRepository.getMemberStreetInfo().onSuccess {
-                _streetInfo.value = it.street
+                _streetInfo.value = it.street.split(" ").take(3)
                 Log.d("street INfo", it.street)
             }
         }
