+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/together/write/WriteTogetherFragment.kt
@@ -229,6 +229,7 @@ class WriteTogetherFragment: BaseFragment<FragmentWriteTogetherBinding>(R.layout
             launch {
                 repeatOnLifecycle(Lifecycle.State.STARTED) {
                     viewModel.dealPlace.collect {
+                        Log.d("dealPlace", it)
                         viewModel.checkDone()
                     }
                 }
